mprotect memory according to brk/tbrk add test for mprotect between brk and tbrk. fix syscall return

+++ runtime/syscall_km.c
@@ -1,6 +1,6 @@
 #include "km_hcalls.h"
 
-int __syscall(long n, long a1, long a2, long a3, long a4, long a5, long a6)
+unsigned long __syscall(long n, long a1, long a2, long a3, long a4, long a5, long a6)
 {
    km_hc_args_t arg;
